requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==19.3.0
basketball-reference-scraper==1.0.1
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
discord.py==1.5.0
fuzzywuzzy==0.18.0
gitdb==4.0.4
GitPython==3.1.1
idna==2.8
idna-ssl==1.1.0
lxml==4.6.3
multidict==4.7.5
nba-api==1.1.9
numpy==1.18.1
pandas==0.25.3
pycparser==2.20
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.22.0
six==1.13.0
smmap==3.0.2
soupsieve==1.9.5
typing-extensions==3.7.4.2
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
youtube-dl==2021.6.6
